# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PayloadProbe Lab API
  version: 0.2.0
  description: >-
    API for managing named XML mock responses. Write endpoints accept non-empty,
    well-formed XML documents or fragments and preserve accepted payload text.
servers:
  - url: https://j4vedskh.github.io/payloadprobe-lab
    description: Documentation site
  - url: http://localhost:8080
    description: Local development
paths:
  /api/responses:
    get:
      summary: List response keys
      responses:
        "200":
          description: Response catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayloadCatalogResponse"
  /api/responses/{key}:
    parameters:
      - $ref: "#/components/parameters/Key"
    get:
      summary: Fetch XML response
      responses:
        "200":
          description: XML response
          content:
            application/xml:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/InvalidKey"
    post:
      summary: Create XML response
      requestBody:
        $ref: "#/components/requestBodies/XmlPayload"
      responses:
        "201":
          $ref: "#/components/responses/MessageResult"
        "409":
          $ref: "#/components/responses/Conflict"
        "400":
          $ref: "#/components/responses/InvalidMutationRequest"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
    put:
      summary: Update XML response
      requestBody:
        $ref: "#/components/requestBodies/XmlPayload"
      responses:
        "200":
          $ref: "#/components/responses/MessageResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/InvalidMutationRequest"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
    delete:
      summary: Delete XML response
      responses:
        "200":
          $ref: "#/components/responses/MessageResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/InvalidKey"
  /api/responses/default:
    post:
      summary: Return default XML sample
      responses:
        "200":
          description: Default XML
          content:
            application/xml:
              schema:
                type: string
  /actuator/health:
    get:
      summary: Health status
      responses:
        "200":
          description: Health status
  /fetch/{key}:
    parameters:
      - $ref: "#/components/parameters/Key"
    get:
      summary: Legacy fetch alias
      responses:
        "200":
          description: XML response or legacy not-found failure XML
          content:
            application/xml:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/InvalidKey"
    post:
      summary: Legacy POST fetch alias
      responses:
        "200":
          description: XML response or legacy not-found failure XML
          content:
            application/xml:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/InvalidKey"
  /add/{key}:
    parameters:
      - $ref: "#/components/parameters/Key"
    post:
      summary: Legacy create alias
      requestBody:
        $ref: "#/components/requestBodies/XmlPayload"
      responses:
        "201":
          $ref: "#/components/responses/MessageResult"
        "409":
          $ref: "#/components/responses/Conflict"
        "400":
          $ref: "#/components/responses/InvalidMutationRequest"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /update/{key}:
    parameters:
      - $ref: "#/components/parameters/Key"
    post:
      summary: Legacy update alias
      requestBody:
        $ref: "#/components/requestBodies/XmlPayload"
      responses:
        "200":
          $ref: "#/components/responses/MessageResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/InvalidMutationRequest"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /delete/{key}:
    parameters:
      - $ref: "#/components/parameters/Key"
    delete:
      summary: Legacy delete alias
      responses:
        "200":
          $ref: "#/components/responses/MessageResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/InvalidKey"
  /fetchAll:
    get:
      summary: Legacy list alias
      responses:
        "200":
          description: Response catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayloadCatalogResponse"
  /default:
    get:
      summary: Legacy default XML alias
      responses:
        "200":
          description: Default XML
          content:
            application/xml:
              schema:
                type: string
    post:
      summary: Legacy default XML alias
      responses:
        "200":
          description: Default XML
          content:
            application/xml:
              schema:
                type: string
  /help:
    get:
      summary: Legacy endpoint catalog
      responses:
        "200":
          description: Endpoint catalog
components:
  requestBodies:
    XmlPayload:
      required: true
      description: >-
        A non-empty, well-formed XML document or fragment containing at least one
        element. The default maximum is 262144 UTF-8 bytes. DTDs and external
        entities are rejected.
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/XmlPayload"
        text/xml:
          schema:
            $ref: "#/components/schemas/XmlPayload"
        text/plain:
          schema:
            $ref: "#/components/schemas/XmlPayload"
  responses:
    MessageResult:
      description: Operation result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    NotFound:
      description: Key not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    Conflict:
      description: Key already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    InvalidKey:
      description: Invalid response key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
          example:
            message: Invalid response key. Use 1-120 letters, numbers, dots, underscores, or hyphens.
    InvalidMutationRequest:
      description: Invalid response key or XML payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
          examples:
            invalidKey:
              value:
                message: Invalid response key. Use 1-120 letters, numbers, dots, underscores, or hyphens.
            blankPayload:
              value:
                message: XML response content must not be blank.
            malformedPayload:
              value:
                message: XML response content must be a well-formed XML document or fragment; DTDs and external entities are not allowed.
            oversizedPayload:
              value:
                message: XML response content exceeds the configured size limit.
    UnsupportedMediaType:
      description: Unsupported request Content-Type
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
          example:
            message: Unsupported Content-Type. Use application/xml, text/xml, or text/plain.
  parameters:
    Key:
      name: key
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,120}$"
  schemas:
    XmlPayload:
      type: string
      minLength: 1
      description: XML document or fragment; the byte limit is enforced after UTF-8 encoding.
      example: <response><status>success</status></response>
    PayloadCatalogResponse:
      type: object
      required:
        - count
        - keys
      properties:
        count:
          type: integer
          example: 3
        keys:
          type: array
          items:
            type: string
          example:
            - openTest
            - paymentSuccess
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Response created for key: paymentSuccess"